Initial implementation of resource API (DM-8283).
Added ResourceConfig class which contains smallest set of fields
describing resource limits used by task. The set will be expanded later
as we know more about it (and if this design is accepted). Added new
method to SuperTask to return the instance of that class corresponding
to the task. Added unit test for these new items.