Add SingleQuantumExecutor class.
This new class implements small piece of cmdLineFwk funcionality which
is responsible for execution of single Quantum. Adds a bunch of
customization points so that new activators can extend default behavior.

No unit test yet for this class, it needs butler and registry mocks,
wouls be easier to make that when we have some simple actual butler for
unit tests.