Make command-line options order-independent (DM-21890)
This commit appears on DM-21889 branch even though the ticket for it is
DM-21890, this is to minimize the number of incompatible merges.

All global/top-level options have been removed and copied to each
sub-command (where they make sense). I have also refactored the code
that makes options into a number of separate methods, this makes it
cleaner and should also prepare us for the next step of refactoring. All
unit tests were updated for new command line interface.