Add support for lsstDebug facility (DM-22504)
The --debug option existed but did not do anyhting, added code that
imports `debug.py` when that option is set. Compared to CmdLineTask in
`pipetask` I do it in two places, once before PreExecInit (in case
someone needs to debug task construction) and once before running each
task. Latter is not needed if `multiprocess` uses 'fork' start method
but it is impossible to guarantee that is always true on all platforms.