Improve printing of config history (DM-27519)
I tried to find a better way to navigate Config structure but in the end
decided that `eval` is the easiest and most natural way to do it.
I did not want to parse the printed attribute representation myself (do
not want to re-implement `eval`) and navigating Config class hierarchy
needs too much effort.