add PipetaskCommand and note to subcommand help output
PipetaskCommand class allows pipetask-specific overrides of
subcommand behaviors.
This adds a comment to the end of every subcommand's help text,
that more options are available using "pipetask --help"

Removes the --log-level option from subcommands.

There was other help text in the epilog that was specific to
prototype development of chained subcommands, and should have
been removed when we did not use that feature. It is now
removed.