Handle special PipelineTask exceptions and rework skip-existing logic.
This redefines the condition for skipping existing quanta to consider
any quantum with a metadata dataset written as complete, even if some
outputs are missing.

That in turn lets us catch NoWorkFound when raised by a PipelineTask
as a signal to just write metadata and move on.  And when we resolve
input datasets before running a quantum, we now call adjustQuantum to
see if we should run the quantum at all (it can say no by raising
NoWorkFound itself).

And because we're now writing metadata even when the task didn't do
anything, this seemed like a good time to start writing the kind of
resource usage that pipe.base.timeMethod provides to the metadata;
this means PipelineTasks do not have to decorate runQuantum themselves
with @timeMethod in order to get timing information written out.

Finally, we catch and handle (when --fail-fast is passed) the other
new pipe_base exceptions by propagating the exit codes they are
associated with so BPS can do what it likes with them.  We also quit
execution immediately when InvalidQuantumError is seen, even when
--fail-fast is not passed.