Fix log-to-butler failures with --skip-existing (DM-31255)
Add special logic to detect cases when we don't want to save log records
to butler. Removes temporary file in that case or when ingest does not
work. When ingest is not supported we now store an empty records object
to make sure that dataset exists for consistency with other cases.