Use temp directory for json log file
This is the only safe way to create a temporary JSON log file
that has default permissions. Temporary files created by tempfile
have restricted positions and need changing but to change it
the entire process needs to have umask changed which can be
dangerous with threading. Safer to create the directory
and write the file normally.