Do not purge datasets when clobbering in execution butler
It's much easier for provenance tracking if the UUID of the
dataset always matches the one generated during initial
creation of the execution butler. When clobbering the
output datasets because of a resubmission of a failed
job the pruneDatasets call was purging the records
and deleting the files. This results in entirely
new UUIDs being used when the new dataset
is put in the butler.  Modify the logic
to not purge if it seems that an execution butler
is being used.