Tweak docstrings to simultaneously satisfy black and flake8.
black likes to put end-triple-quotes for docstrings on the same line
they started on whenever possible, and doesn't care if that puts the
line length over 79.  Flake8 cares if a doc line goes over 79 chars.

We don't care, but there is apparently no set of configurations to
make them stop fighting, so I've just changed the actual docstrings to
avoid the conflict.  Whenever possible, I've tried to make them
better, but sometimes it's just a neutral rewording.