Continue execution after task failure (DM-33339)
In a single-process mode the MPGraphExecutor will try to execute remaining
tasks after one of the tasks fails (skipping the tasks that depend on
outputs of failed tasks) instead of exiting on an exception. This needed
some changes in exception handling, instead of passing an original
exeption raised by a task, it now raises MPGraphExecutorError, this is
consistent with multi-process case. Task timeouts are not handled in
single-process case.