add --pdb option to drop into debugger on exception
When we get an exception, sometimes it's not enough to just
see the backtrace, but we also want to get a debugger prompt
at the point of the exception.

Replaced the --do-raise option (apparently unused, and often
associated with getting a debugger with Gen2) with --pdb
(well-known and appreciated option from pytest), which causes
MPGraphExecutor._executeQuantaInProcess to drop into pdb.