remove unused, untested function
There is a split_commas function in daf_butler that can be used.
It can be extended to do type conversion (as this one did, to
int), or we could write a helper to convert a list of Any items
to a list of int items that could be chained with split_commas.