executionButler: add option to transfer files
In some instances, it is helpful/important to transfer files
into the execution butler (e.g., original butler datastore is
on a slow disk). Added option:
    pipetask qgraph ... --target-datastore-root=<root>
to allow transfer of files into a new datastore. Also added
a --transfer option for fine-grained control of the transfer
mode, which defaults to "copy" if there's a new datastore.