Try to set the datastore cache directory explicitly
We have to do it early to ensure that in fork mode the
cache is defined before the initial output Butler is created
(fork will not pickle it).

Spawn will create a new pickled butler that will inherit the environment
and in that scenario the setting of the environment variable
can reliably happen in MPGraphExecutor.