Force the pipeline to run with the same output run as used in the graph
Without this we were on occasion having the situation where the
graph was being created with timestamp N and the pipeline was
being created with an output run of timestamp N+1 as we crossed
a second boundary. With unresolved refs this transparently resolved
itself because the datasets would end up in the N+1 output run
and nothing would notice that output run N was empty. Once resolved
refs from the graph are believed, this breaks because the graph
output run isn't even necessarily defined in this output butler.