Refresh pipeline directed graph colors
The current pipeline directed graph nodes use dark gray as their
background color. It has been reported that it is difficult to read the
black text on the dark gray background. This ticket updates the directed
graph color scheme with an aim towards making node text easier to read.
In the process of exploring what color schemes would be optimal to
satisfy the aim of this ticket, it emerged that making use of the Rubin
visual identity colors may be desirable. This will help to make LSST
pipeline graphs more instantly recognizable as Rubin-associated
products. Colors: https://rubin.canto.com/g/RubinVisualIdentity