Extract instrument from pipeline when building graph.
This was previously done by logic in GraphBuilder.makeGraph, but only
when passing in a Pipeline object rather than the sequence of TaskDef
that represents an "expanded" pipeline, since the TaskDef form can't
carry around the instrument.  Now that we're passing in a sequence of
TaskDefs (to simplify the mock/no-mock logic branching) we need to do
it here.

Eventually, we'll want a representation of the expanded pipeline that
can remember the instrument, and the PipelineGraph class being added
on DM-33027 will play that role, but this probably won't be updated
until DM-38498, which is where we'll stop using TaskDef in graph
generation.