Fix bugs in test pipeline task storage conversions.
All of these tasks (modeled on lsst.pipe.base.tests.NoDimensionsTask,
which is also in use here) were using their 'outputSC' config field
to change the pytype they returned without adjusting their connection
definition accordingly, and in one case that meant testing for the
wrong behavior.

I've also changed the default value for 'outputSC' from 'dict' to
'StructuredDataDict' since it's supposed to be a storage class.  Given
that we were only testing to see if it was something else (like
'TaskMetadata'), this didn't matter, but it was confusing.