Restore previous pipetask run-qbb behavior.
It's not clear whether previous authors were aware that passing
clobberOutputs=True and skipExisting=True in CmdLineFwk was previously
doing nothing because SingleQuantumExecution executor shortcutted
checkExistingOutputs when there was only a limited butler.

But while it looks safe from a correctness standpoint to just drop
that shortcut and let those options work as intended, that could mean
a lot of new file existence checks in BPS jobs, so it's safer for this
ticket to focus on enabling clobbering and skip-existing for other
LimitedButlers while leaving run-qbb behavior strictly unchanged.