Improve exception handling in quantum executors (DM-45340)
This should solve an issue with exceptions happening at unpredictable
locations in a SingleQuantumExecutor class. MPGraphExecutor in a single-process
mode should also behave reasonably for random exceptions. Multi-process mode
has more complex internal state. If a client of the code catches an exception
and tries to reuse the executor, the state may be inconsistent, depending
on where the exception happens.