Fix dimension universe instantiation in subprocesses (DM-45894)
With run-qbb when dimension universe is not the default one, we need
to make sure that that universe is instantiated from a configuration
in a subprocess before unpickling anything that requires that universe.
The fix uses butler factory to do that, using its unpickle method,
and there is some complicated logic in executors to make sure that
unpickling of that factory happens first thing in a subprocess.
There may be an easier way to enforce that but it would require
interface changes for MPGraphExecutor or SingleQuantumExecutor.